code refactor in input number component
